Manage firefox with home-manager
Allows home-manager to manage prefered settings, extensions, and ui of
Firefox. Extensions are packaged by an NUR project, but does not include
preferred theme (Matte (Green)), some day may package and include that
myself. There is a disabled hack to hide the "this time, search with"
prompt from the URL bar, but I cannot find a long term solution to fully
disable that functionality. Maybe build Firefox without it?